history dialog view leaves memory leak fix and description no escape bug fix and type modify

+++ app/src/main/java/com/puutaro/commandclick/activity_lib/InitManager.kt
@@ -202,7 +202,7 @@ class InitManager(
             })
             .show()
         alertDialog.getButton(DialogInterface.BUTTON_POSITIVE).setTextColor(
-            activity.getColor(android.R.color.black) as Int
+            activity.getColor(android.R.color.black)
         )
     }
